- readability improvements to the ServiceMap component - only show the update map button toast after last request loads
